chore: preparing to remove AdvancedSearch references from tao-core
