+++ models/classes/resources/ResourceWatcher.php
@@ -21,22 +21,13 @@
 
 namespace oat\tao\model\resources;
 
-use common_http_Request;
-use core_kernel_classes_Class;
 use core_kernel_classes_Resource;
 use oat\generis\model\data\event\ResourceCreated;
 use oat\generis\model\data\event\ResourceDeleted;
 use oat\generis\model\data\event\ResourceUpdated;
 use oat\generis\model\OntologyAwareTrait;
-use oat\generis\model\OntologyRdfs;
 use oat\oatbox\service\ConfigurableService;
-use oat\tao\model\AdvancedSearch\AdvancedSearchChecker;
-use oat\tao\model\search\index\IndexUpdaterInterface;
-use oat\tao\model\search\Search;
-use oat\tao\model\search\tasks\UpdateClassInIndex;
-use oat\tao\model\search\tasks\UpdateResourceInIndex;
 use oat\tao\model\TaoOntology;
-use oat\tao\model\taskQueue\QueueDispatcherInterface;
 
 /**
  * @author Aleksej Tikhanovich, <aleksej@taotesting.com>
@@ -63,9 +54,6 @@ class ResourceWatcher extends ConfigurableService
         $resource->editPropertyValues($property, $now);
 
         $this->getLogger()->debug('triggering index update on resourceCreated event');
-
-        $taskMessage = __('Adding search index for created resource');
-        $this->createResourceIndexingTask($resource, $taskMessage);
     }
 
     /**
@@ -85,9 +73,6 @@ class ResourceWatcher extends ConfigurableService
         if ($updatedAt === null || ($now - $updatedAt) > $threshold) {
             $this->getLogger()->debug('triggering index update on resourceUpdated event');
 
-            $taskMessage = __('Adding/updating search index for updated resource');
-            $this->createResourceIndexingTask($resource, $taskMessage);
-
             $property = $this->getProperty(TaoOntology::PROPERTY_UPDATED_AT);
             $this->updatedAtCache[$resource->getUri()] = $now;
             $resource->editPropertyValues($property, $now);
@@ -96,15 +81,7 @@ class ResourceWatcher extends ConfigurableService
 
     public function catchDeletedResourceEvent(ResourceDeleted $event): void
     {
-        $searchService = $this->getServiceLocator()->get(Search::SERVICE_ID);
-        try {
-            $searchService->remove($event->getId());
-        } catch (\Exception $e) {
-            $message = $e->getMessage();
-            $this->getLogger()->error(
-                sprintf("Error delete index document for %s with message %s", $event->getId(), $message)
-            );
-        }
+        $this->getLogger()->debug('Method' . __METHOD__ . ' shall not be called anymore');
     }
 
      /**
@@ -114,91 +91,18 @@ class ResourceWatcher extends ConfigurableService
     public function getUpdatedAt(core_kernel_classes_Resource $resource)
     {
         if (isset($this->updatedAtCache[$resource->getUri()])) {
-            $updatedAt = $this->updatedAtCache[$resource->getUri()];
-        } else {
-            $property = $this->getProperty(TaoOntology::PROPERTY_UPDATED_AT);
-            $updatedAt = $resource->getOnePropertyValue($property);
-            if ($updatedAt && $updatedAt instanceof \core_kernel_classes_Literal) {
-                $updatedAt = (int) $updatedAt->literal;
-            }
-            $this->updatedAtCache[$resource->getUri()] = $updatedAt;
+            return $this->updatedAtCache[$resource->getUri()];
         }
+
+        $property = $this->getProperty(TaoOntology::PROPERTY_UPDATED_AT);
+        $updatedAt = $resource->getOnePropertyValue($property);
+
+        if ($updatedAt && $updatedAt instanceof \core_kernel_classes_Literal) {
+            $updatedAt = (int) $updatedAt->literal;
+        }
+
+        $this->updatedAtCache[$resource->getUri()] = $updatedAt;
+
         return $updatedAt;
-    }
-
-    /**
-     * Creates a task in the task queue to index/re-index created/updated resource
-     */
-    private function createResourceIndexingTask(core_kernel_classes_Resource $resource, string $message): void
-    {
-        if ($this->getServiceLocator()->get(AdvancedSearchChecker::class)->isEnabled()) {
-            /** @var QueueDispatcherInterface $queueDispatcher */
-            $queueDispatcher = $this->getServiceLocator()->get(QueueDispatcherInterface::SERVICE_ID);
-
-            if ($this->hasClassSupport($resource) && !$this->ignoreEditIemClassUpdates()) {
-                $queueDispatcher->createTask(new UpdateClassInIndex(), [$resource->getUri()], $message);
-                return;
-            }
-
-            if ($this->hasResourceSupport($resource)) {
-                $queueDispatcher->createTask(new UpdateResourceInIndex(), [$resource->getUri()], $message);
-
-                return;
-            }
-        }
-    }
-
-    private function hasResourceSupport(core_kernel_classes_Resource $resource): bool
-    {
-        $resourceTypeIds = array_map(
-            function (core_kernel_classes_Class $resourceType): string {
-                return $resourceType->getUri();
-            },
-            $resource->getTypes()
-        );
-
-        $checkedResourceTypes = [OntologyRdfs::RDFS_RESOURCE, TaoOntology::CLASS_URI_OBJECT];
-        $resourceTypeIds = array_diff($resourceTypeIds, [OntologyRdfs::RDFS_RESOURCE, TaoOntology::CLASS_URI_OBJECT]);
-
-        while (!empty($resourceTypeIds)) {
-            $classUri = array_pop($resourceTypeIds);
-
-            $hasClassSupport = $this->getServiceLocator()
-                ->get(IndexUpdaterInterface::SERVICE_ID)
-                ->hasClassSupport(
-                    $classUri
-                );
-
-            if ($hasClassSupport) {
-                return true;
-            }
-
-            $class = $this->getClass($classUri);
-
-            foreach ($class->getParentClasses() as $parent) {
-                if (!in_array($parent->getUri(), $checkedResourceTypes)) {
-                    $resourceTypeIds[] = $parent->getUri();
-                }
-            }
-            $checkedResourceTypes[] = $class->getUri();
-        }
-
-        return false;
-    }
-
-    private function hasClassSupport(core_kernel_classes_Resource $resource): bool
-    {
-        return $resource instanceof core_kernel_classes_Class;
-    }
-
-    private function ignoreEditIemClassUpdates(): bool
-    {
-        try {
-            $url = parse_url(common_http_Request::currentRequest()->getUrl());
-        } catch (\common_exception_Error $e) {
-            return false;
-        }
-
-        return isset($url['path']) && $url['path'] === '/taoItems/Items/editItemClass';
     }
 }
