+++ test/unit/models/classes/resources/ResourceWatcherTest.php
@@ -43,6 +43,7 @@ use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\MockObject\Rule\InvocationOrder;
 use Psr\Log\LoggerInterface;
 
+//@TODO Fix this test
 class ResourceWatcherTest extends TestCase
 {
     /** @var ResourceWatcher|MockObject */
